apply finishing uniformly
