+++ src/dataflow-types/Cargo.toml
@@ -11,6 +11,7 @@ path = "lib.rs"
 [dependencies]
 comm = { path = "../comm" }
 expr = { path = "../expr" }
+pdqselect = "0.1.0"
 repr = { path = "../repr" }
 serde = { version = "1.0", features = ["derive"] }
 url = "1.7.2"
